+++ src/main/java/roomescape/service/dto/response/reservation/ReservationResponse.java
@@ -2,7 +2,7 @@ package roomescape.service.dto.response.reservation;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
-import roomescape.service.dto.response.ReservationTimeResponse;
+import roomescape.service.dto.response.reservationTime.ReservationTimeResponse;
 import roomescape.service.dto.response.ThemeResponse;
 import roomescape.service.dto.response.member.MemberIdAndNameResponse;
