+++ src/main/java/roomescape/service/dto/response/reservationTime/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.response;
+package roomescape.service.dto.response.reservationTime;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
