+++ src/main/resources/static/js/time.js
@@ -20,7 +20,7 @@ function render(data) {
     const tableBody = document.getElementById('table-body');
     tableBody.innerHTML = '';
 
-    data.forEach(item => {
+    data.reservationTimes.forEach(item => {
         const row = tableBody.insertRow();
 
         cellFields.forEach((field, index) => {
